feat(markdown): Do not trim last two in markdown
The trim-trailing-whitspace feature would indiscriminately trim the
trailing whitespace in any file type.  This patch adjusts it so that
for markdown (injected or top level) if there's two trailing whitespace
characters then they are retained (single trailing is dropped, and
more than two is trimmed to two) to maintain explicit line breaks.

Signed-off-by: Daniel Silverstone <dsilvers@digital-scurf.org>